Moved briefing-content to components, added checkmark remove functionality
