+++ src/components/briefing-content.tsx
